- Update Github actions to use current actions to avoid Node 16 deprecation notice. - Add validation checks to module uninstall script.

+++ app/module/uninstall.sh
@@ -12,10 +12,18 @@ APP_EXTERNAL_DATA_DIR="${EXTERNAL_DATA_DIR}${APP_PACKAGE}"
 find /data/dalvik-cache -iname '*pixelupdater*' | xargs rm -f
 find /data/system/package_cache -iname '*pixelupdater*' | xargs rm -f
 
-find /data/ota_package -user $APP_UID -exec rm -rf {} \;
-
-rm -rf $APP_DATA_DIR
-if [ -d "/data/data/$APP_PACKAGE" ]; then
-    rm -rf /data/data/$APP_PACKAGEr
+if [ -n "$APP_UID" ]; then
+    find /data/ota_package -user $APP_UID -exec rm -rf {} \;
 fi
-rm -rf $APP_EXTERNAL_DATA_DIR
+
+if [ -n "$APP_DATA_DIR" ]; then
+    rm -rf $APP_DATA_DIR
+fi
+
+if [ -d "/data/data/$APP_PACKAGE" ]; then
+    rm -rf /data/data/$APP_PACKAGE
+fi
+
+if [ -d $APP_EXTERNAL_DATA_DIR ]; then
+    rm -rf $APP_EXTERNAL_DATA_DIR
+fi
